chore: Consolidate existing raster unit tests
